requirements.txt
matplotlib==3.9.2
numpy==1.26.4
pillow==10.4.0
PyYAML==6.0.2
streamlit==1.38.0
tensorboard==2.18.0
tensorboardX==2.6.2.2
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
